mac-sys-utils:  Fix get_address_mask() on 68000
Our broken word-wide access of CPUFlag yielded an Address Error on 68000
(and 68010) processors.  On a 68020, the access would successfully get a
value of at least 512, with the same result as if it had been read as 2.

+++ mac/libs/mac-sys-utils/mac_sys/get_address_mask.hh
@@ -16,9 +16,9 @@
 namespace mac {
 namespace sys {
 
-#define CPUFlag     (*(unsigned short*) 0x012f)
-#define SysVersion  (*(unsigned short*) 0x015a)
-#define Lo3Bytes    (*(unsigned long *) 0x031a)
+#define CPUFlag     (*(UInt8 *) 0x012f)
+#define SysVersion  (*(UInt16*) 0x015a)
+#define Lo3Bytes    (*(UInt32*) 0x031a)
 
 inline
 unsigned long get_address_mask()
